ProyectoFinalTests/BLL/ClientesBLLTests.cs
using Microsoft.VisualStudio.TestTools.UnitTesting;
using ProyectoFinal.BLL;
using ProyectoFinal.Entidades;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace ProyectoFinal.BLL.Tests
{
    [TestClass()]
    public class ClientesBLLTests
    {
        [TestMethod()]
        public void GuardarTest()
        {
            Cliente cliente = new Cliente
            {
                FechaNacimiento = DateTime.Now,
                Nombres = "Victor",
                Apellidos = "Guichardo",
                Direccion = "Calle 3 #3 EnsancheDuarte",
                Telefono = "809-541-9654",
                Celular = "829-985-4741",
                Email = "guichardo@gmail.com"
            };

            bool paso = ClientesBLL.Guardar(cliente);
            Assert.AreEqual(true, paso);
        }

        [TestMethod()]
        public void ModificarTest()
        {
            Assert.Fail();
        }

        [TestMethod()]
        public void EliminarTest()
        {
            Assert.Fail();
        }

        [TestMethod()]
        public void BuscarTest()
        {
            Assert.Fail();
        }

        [TestMethod()]
        public void GetListTest()
        {
            Assert.Fail();
        }

        [TestMethod()]
        public void GetListTest1()
        {
            Assert.Fail();
        }
    }
}